requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==2.0.0
Django==5.2.5
gunicorn==21.2.0
idna==3.10
psycopg2-binary==2.9.10
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.5.0
psutil==7.1.3
aiohttp==3.13.2